Consolidate download service
