+++ spec/fixtures/vcr/log_level_data_service_download.yml
@@ -1,0 +1,105 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api-test.appnexus.com/auth
+    body:
+      encoding: UTF-8
+      string: '{"auth":{"username":"<APPNEXUS_USERNAME>","password":"<APPNEXUS_PASSWORD>"}}'
+    headers:
+      User-Agent:
+      - Faraday v0.11.0
+      Content-Type:
+      - application/json
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Cache-Control:
+      - no-store, no-cache, must-revalidate, post-check=0, pre-check=0;no-cache
+      Content-Length:
+      - '698'
+      Content-Type:
+      - application/json
+      Date:
+      - Mon, 13 Feb 2017 10:00:37 GMT
+      Expires:
+      - Thu, 19 Nov 1981 08:52:00 GMT
+      Pragma:
+      - no-cache
+      Server:
+      - Apache
+      Set-Cookie:
+      - HBFAPI_SESSID=hbapi%3A171678%3A58a183c65f9a3%3Anym2; Path=/
+      X-Route:
+      - "/auth"
+      X-Route-Time:
+      - '1375'
+    body:
+      encoding: UTF-8
+      string: '{"response":{"status":"OK","token":"hbapi:171678:58a183c65f9a3:nym2","dbg_info":{"instance":"01.hbapi.test104169.nym2","slave_hit":false,"db":"master","parent_dbg_info":{"instance":"01.api.test104169.nym2","slave_hit":false,"db":"master","awesomesauce_cache_used":false,"count_cache_used":false,"uuid":"66f42891-11b1-505a-9099-ff839b2e0875","warnings":[],"time":792.37198829651,"start_microtime":1486980037.2725,"version":"1.17","output_term":"not_found"},"awesomesauce_cache_used":false,"count_cache_used":false,"uuid":"be05b49d-5dbd-5334-a9da-18e6edee3563","warnings":[],"time":1363.0979061127,"start_microtime":1486980037.247,"version":"1.17.202","slave_lag":1461580,"output_term":"not_found"}}}'
+    http_version: 
+  recorded_at: Mon, 13 Feb 2017 10:00:36 GMT
+- request:
+    method: get
+    uri: https://api-test.appnexus.com/siphon-download?num_elements=100&siphon_name=standard_feed&start_element=0&updated_since=2017_02_13_00
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - Faraday v0.11.0
+      Authorization:
+      - hbapi:171678:58a183c65f9a3:nym2
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Cache-Control:
+      - no-store, no-cache, must-revalidate, post-check=0, pre-check=0;no-cache
+      Content-Length:
+      - '787'
+      Content-Type:
+      - application/json
+      Date:
+      - Mon, 13 Feb 2017 10:00:39 GMT
+      Expires:
+      - Thu, 19 Nov 1981 08:52:00 GMT
+      Pragma:
+      - no-cache
+      Server:
+      - Apache
+      Set-Cookie:
+      - HBFAPI_SESSID=hbapi%3A171678%3A58a183c65f9a3%3Anym2; Path=/
+      X-Count-Read:
+      - user:8,member:5,serviceHostUser:1,serviceHostMember:1,hostUser:5,hostMember:5,ip:0
+      X-Count-Write:
+      - user:0,member:0,serviceHostUser:0,serviceHostMember:0,hostUser:0,hostMember:0,ip:0
+      X-Ratelimit-Read:
+      - '1000'
+      X-Ratelimit-Service:
+      - 1000-Default
+      X-Ratelimit-System:
+      - 1000-Default
+      X-Ratelimit-Write:
+      - '1000'
+      X-Route:
+      - "/siphon-download"
+      X-Route-Time:
+      - '22'
+    body:
+      encoding: UTF-8
+      string: '{"response":{"error_id":"SYNTAX","error":"hour is required","error_description":null,"service":"siphon-download","method":"GET","error_code":null,"dbg_info":{"instance":"01.hbapi.test104169.nym2","slave_hit":false,"db":"master","user::reads":8,"user::read_limit":1000,"user::read_limit_seconds":60,"user::writes":0,"user::write_limit":1000,"user::write_limit_seconds":60,"reads":5,"read_limit":1000,"read_limit_seconds":60,"writes":0,"write_limit":1000,"write_limit_seconds":60,"awesomesauce_cache_used":false,"count_cache_used":false,"uuid":"15316bb9-6d1e-53cc-91a3-11de3e32fe49","warnings":[],"time":14.532089233398,"start_microtime":1486980039.1295,"version":"1.17.202","slave_miss":"no_timestamp","slave_lag":1461581,"member_last_modified_age":1486980039,"output_term":"not_found"}}}'
+    http_version: 
+  recorded_at: Mon, 13 Feb 2017 10:00:37 GMT
+recorded_with: VCR 3.0.3
