+++ spec/integration/placement_spec.rb
@@ -1,7 +1,6 @@
 require 'spec_helper'
 
 describe AppnexusApi::PlacementService do
-
   include_context 'with a publisher'
 
   let(:placement_service) { described_class.new(connection) }
